Moved the integration tests for taquito-local-forging package to integration tests folder
