+++ integration-tests/local-forging.spec.ts
@@ -1,0 +1,34 @@
+import { CONFIGS } from "./config";
+import { commonCases, edoCases } from '../packages/taquito-local-forging/test/allTestsCases';
+import { localForger } from '@taquito/local-forging'
+import { Protocols, TezosToolkit } from "@taquito/taquito";
+
+CONFIGS().forEach(({ rpc, protocol }) => {
+    const Tezos = new TezosToolkit(rpc);
+    const edonet = (protocol === Protocols.PtEdo2Zk) ? test : test.skip;
+
+    describe(`Test local forger: ${rpc}`, () => {
+
+        commonCases.forEach(({ name, operation, expected }) => {
+
+            it(`Should give the same result as when forging with the rpc: ${name} (${rpc})`, async done => {
+                const result = await localForger.forge(operation);
+                const rpcResult = await Tezos.rpc.forgeOperations(operation);
+                expect(result).toEqual(rpcResult);
+                expect(await localForger.parse(result)).toEqual(expected || operation);
+                done();
+            });
+        });
+
+         edoCases.forEach(({ name, operation, expected }) => {
+
+            edonet(`Should give the same result as when forging with the rpc: ${name} (${rpc})`, async done => {
+                const result = await localForger.forge(operation);
+                const rpcResult = await Tezos.rpc.forgeOperations(operation);
+                expect(result).toEqual(rpcResult);
+                expect(await localForger.parse(result)).toEqual(expected || operation);
+                done();
+            });
+        }); 
+    });
+})
