+++ packages/taquito-local-forging/test/taquito-local-forging.spec.ts
@@ -1,875 +1,18 @@
-import { TezosToolkit } from '@taquito/taquito';
-import { opMappingReverse } from '../src/constants';
 import { localForger } from '../src/taquito-local-forging';
-import { 
-  rpcContractResponse, 
-  rpcContractResponse2, 
-  rpcContractResponse3, 
-  rpcContractResponse4, 
-  rpcContractResponse5, 
-  rpcContractResponse6, 
-  rpcContractResponse7, 
-  rpcContractResponse8, 
-  rpcContractResponse9 
-} from './data/code_with_sapling';
-import { ticketCode, 
-  ticketCode2, 
-  ticketCode3, 
-  ticketCode4, 
-  ticketStorage, 
-  ticketStorage2, 
-  ticketStorage3, 
-  ticketStorage4 
-} from './data/code_with_ticket';
-import { genericCode, genericStorage } from './data/generic_contract';
-import { tokenBigmapCode, tokenBigmapStorage } from './data/token_big_map';
-import { noAnnotCode, noAnnotInit } from './data/token_without_annotations';
-import { voteInitSample, voteSample } from './data/vote_contract';
+import { commonCases, edoCases } from './allTestsCases';
 
-const integrationTest = process.env.RUN_INTEGRATION ? test : test.skip;
-
-interface TestCase {
-  name: string;
-  operation: any;
-  expected?: {};
-}
-
-const cases: TestCase[] = [
-  {
-    name: 'Delegation',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'delegation',
-          delegate: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Reveal',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'reveal',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          public_key: 'edpkvS5QFv7KRGfa3b87gg9DBpxSm3NpSwnjhUjNBQrRUUR66F7C9g',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Ballot',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'ballot',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          period: -300,
-          ballot: 'yay',
-          proposal: 'PsCARTHAGazKbHtnKfLzQg3kms52kSRpgnDY982a9oYsSXRLQEb',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Endorsement',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'endorsement',
-          level: -300,
-        },
-      ],
-    },
-  },
-  {
-    name: 'Seed nonce revelation',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'seed_nonce_revelation',
-          level: 25550,
-          nonce: new Array(32).fill('ff').join(''),
-        },
-      ],
-    },
-  },
-  {
-    name: 'Proposals',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'proposals',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          period: 25550,
-          proposals: [
-            'PsCARTHAGazKbHtnKfLzQg3kms52kSRpgnDY982a9oYsSXRLQEb',
-            'PsCARTHAGazKbHtnKfLzQg3kms52kSRpgnDY982a9oYsSXRLQEb',
-          ],
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          destination: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with parameter',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'do',
-            value: { bytes: '0202' },
-          },
-          destination: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with default entrypoint',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'default',
-            value: { prim: 'Pair', args: [{ int: '2' }, { string: 'hello' }] },
-          },
-          destination: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with maximum length entrypoint',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'Tps0RV2UISBvTV6m2z16VnfCVnN5dzX',
-            value: { prim: 'Pair', args: [{ int: '2' }, { string: 'hello' }] },
-          },
-          destination: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with non ascii entrypoint and string',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'entrypoint©Ͻ',
-            value: { string: 'Copyright ©Ͻ' },
-          },
-          destination: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with default entrypoint and unit parameter',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'default',
-            value: { prim: 'Unit' },
-          },
-          destination: 'KT1JHqHQdHSgWBKo6H4UfG8dw3JnZSyjGkHA',
-          amount: '1000',
-        },
-      ],
-    },
-    expected: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          destination: 'KT1JHqHQdHSgWBKo6H4UfG8dw3JnZSyjGkHA',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with set_delegate entrypoint',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'set_delegate',
-            value: { string: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn' },
-          },
-          destination: 'KT1JHqHQdHSgWBKo6H4UfG8dw3JnZSyjGkHA',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with remove_delegate entrypoint',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'remove_delegate',
-            value: { string: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn' },
-          },
-          destination: 'KT1JHqHQdHSgWBKo6H4UfG8dw3JnZSyjGkHA',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with root entrypoint',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'root',
-            value: { string: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn' },
-          },
-          destination: 'KT1JHqHQdHSgWBKo6H4UfG8dw3JnZSyjGkHA',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with do entrypoint',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'do',
-            value: { string: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn' },
-          },
-          destination: 'KT1JHqHQdHSgWBKo6H4UfG8dw3JnZSyjGkHA',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with do entrypoint and unit',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'do',
-            value: { prim: 'Unit' },
-          },
-          destination: 'KT1JHqHQdHSgWBKo6H4UfG8dw3JnZSyjGkHA',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with custom entrypoint',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'main',
-            value: {
-              prim: 'Pair',
-              args: [
-                { prim: 'Pair', args: [{ bytes: '0202' }, { int: '202' }] },
-                { string: 'hello' },
-              ],
-            },
-          },
-          destination: 'KT1HPaJE1QNtuiYPgMAGhzTrs446K9wptmsR',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Transaction with default entrypoint and annotation',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'default',
-            value: {
-              prim: 'Pair',
-              args: [{ bytes: '0202' }, { string: 'hello' }],
-              annots: ['%test'],
-            },
-          },
-          destination: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination vote example',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: voteSample,
-            storage: voteInitSample,
-          },
-        },
-      ],
-    },
-  },
-  ...Object.keys(opMappingReverse).map(op => {
-    return {
-      name: `Origination operation (${op})`,
-      operation: {
-        branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-        contents: [
-          {
-            kind: 'origination',
-            counter: '1',
-            source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-            fee: '10000',
-            gas_limit: '10',
-            storage_limit: '10',
-            balance: '0',
-            script: {
-              code: genericCode(op),
-              storage: genericStorage,
-            },
-          },
-        ],
-      },
-    };
-  }),
-  {
-    name: 'Origination with bigmap',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: tokenBigmapCode,
-            storage: tokenBigmapStorage,
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with no annot',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: noAnnotCode,
-            storage: noAnnotInit('tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn'),
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Multiple transaction at once',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'reveal',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          public_key: 'edpkvS5QFv7KRGfa3b87gg9DBpxSm3NpSwnjhUjNBQrRUUR66F7C9g',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-        },
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: noAnnotCode,
-            storage: noAnnotInit('tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn'),
-          },
-        },
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          parameters: {
-            entrypoint: 'default',
-            value: {
-              prim: 'Pair',
-              args: [{ bytes: '0202' }, { string: 'hello' }],
-              annots: ['%test'],
-            },
-          },
-          destination: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          amount: '1000',
-        },
-        {
-          kind: 'transaction',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          destination: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          amount: '1000',
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with ticket 1',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: ticketCode,
-            storage: ticketStorage,
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with ticket 2',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: ticketCode2,
-            storage: ticketStorage2,
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with ticket 3',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: ticketCode3,
-            storage: ticketStorage3,
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with ticket 4',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: ticketCode4,
-            storage: ticketStorage4,
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with sapling 1',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse.script.code,
-            storage: [],
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with sapling 2',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse2.script.code,
-            storage: { prim: 'Pair', args: [{ int: '0' }, [], []] },
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with sapling 3',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse3.script.code,
-            storage: [],
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with sapling 4',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse4.script.code,
-            storage: { prim: 'Pair', args: [[], []] },
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with sapling 5',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse5.script.code,
-            storage: { prim: 'Unit' },
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with sapling 6',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse6.script.code,
-            storage: { prim: 'Unit' },
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with sapling 7',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse7.script.code,
-            storage: { prim: 'None' },
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'Origination with sapling 8',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse8.script.code,
-            storage: [],
-          },
-        },
-      ],
-    },
-  },
-  {
-    name: 'A test with comb pairs',
-    operation: {
-      branch: 'BLzyjjHKEKMULtvkpSHxuZxx6ei6fpntH2BTkYZiLgs8zLVstvX',
-      contents: [
-        {
-          kind: 'origination',
-          counter: '1',
-          source: 'tz1QZ6KY7d3BuZDT1d19dUxoQrtFPN2QJ3hn',
-          fee: '10000',
-          gas_limit: '10',
-          storage_limit: '10',
-          balance: '0',
-          script: {
-            code: rpcContractResponse9.script.code,
-            storage: [],
-          },
-        },
-      ],
-    },
-  },
-];
-
-cases.forEach(({ name, operation, expected }) => {
-  test(`Test: ${name}`, async done => {
+commonCases.forEach(({ name, operation, expected }) => {
+  test(`Common test: ${name}`, async done => {
     const result = await localForger.forge(operation);
     expect(await localForger.parse(result)).toEqual(expected || operation);
     done();
   });
+});
 
-  ['https://api.tez.ie/rpc/delphinet', 'https://api.tez.ie/rpc/edonet'].forEach(rpc => {
-    integrationTest(`Integration test: ${name} (${rpc})`, async done => {
-      const Tezos = new TezosToolkit(rpc);
-      Tezos.setProvider({ rpc });
-      const result = await localForger.forge(operation);
-
-      const rpcResult = await Tezos.rpc.forgeOperations(operation);
-
-      expect(result).toEqual(rpcResult);
-      expect(await localForger.parse(result)).toEqual(expected || operation);
-      done();
-    });
+edoCases.forEach(({ name, operation, expected }) => {
+  test(`Edo test: ${name}`, async done => {
+    const result = await localForger.forge(operation);
+    expect(await localForger.parse(result)).toEqual(expected || operation);
+    done();
   });
 });
